courseapisv1/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
pillow==11.0.0
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
setuptools==60.2.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
wheel==0.37.1
